refactor(sdk): extract RFQ + notifications + network to _client/_rfq.py
PR 4/N of the SDK refactor split (see #702 WebSocket, #706 Guardian,
#711 Discovery for the mixin template). Moves the RFQ workflow,
single-use transaction-token issuance, per-agent notifications and
the network-onboarding classmethod out of cullis_sdk/client.py into a
dedicated mixin under cullis_sdk/_client/.

What moves:

- cullis_sdk/_client/_rfq.py — new module exposing _RfqMixin with:
    create_rfq, respond_to_rfq, get_rfq           (# ── RFQ ──)
    request_transaction_token                     (# ── Transaction Tokens ──)
    get_notifications                             (# ── Notifications ──)
    join_network                                  (# ── Onboarding ──)
  Plus the scoped import of RfqResult from cullis_sdk.types.

- cullis_sdk/client.py:
  * CullisClient now inherits
    (_DiscoveryMixin, _GuardianMixin, _RfqMixin, _WebSocketMixin)
    in alphabetical order (D < G < R < W).
  * Six methods, their four section comments and the now-unused
    ``RfqResult`` import in client.py are removed.

- cullis_sdk/__init__.py — unchanged. ``RfqResult`` and ``RfqQuote``
  continue to be re-exported via ``from cullis_sdk.types import …``,
  same path as before.

``join_network`` needs the module-level ``_check_insecure_tls`` helper
that lives in cullis_sdk.client. Top-level import would create a
load-order cycle (client.py imports _rfq before the helper is
defined), so the call site uses a lazy ``from cullis_sdk.client
import _check_insecure_tls`` inside the classmethod — same pattern
the SDK already uses for ``verify_ticket`` inside the Guardian path.

No logic change — pure code movement.
