+++ cullis_sdk/_client/_rfq.py
@@ -1,0 +1,147 @@
+"""RFQ + transaction tokens + notifications + network onboarding
+extracted from :mod:`cullis_sdk.client`.
+
+Single public symbol:
+
+* :class:`_RfqMixin` — mixin folded into ``CullisClient`` that exposes
+  the RFQ workflow (``create_rfq`` / ``respond_to_rfq`` / ``get_rfq``),
+  one-shot transaction token issuance (``request_transaction_token``),
+  per-agent notification fetch (``get_notifications``), and the
+  classmethod that bootstraps a new org onto the network
+  (``join_network``).
+
+Movement only — no behavior change. The mixin assumes the host class
+exposes ``_authed_request`` and ``_label``. ``join_network`` is a
+classmethod that does not touch instance state.
+"""
+from __future__ import annotations
+
+from typing import Any
+
+import httpx
+
+from cullis_sdk.types import RfqResult
+
+
+class _RfqMixin:
+    """RFQ workflow, transaction tokens, notifications and network
+    onboarding on ``CullisClient``."""
+
+    # ── RFQ ─────────────────────────────────────────────────────────
+
+    def create_rfq(
+        self,
+        capability_filter: list[str],
+        payload: dict,
+        timeout_seconds: int = 30,
+    ) -> RfqResult:
+        """Broadcast an RFQ to agents matching capability filter."""
+        path = "/v1/broker/rfq"
+        resp = self._authed_request("POST", path, json={
+            "capability_filter": capability_filter,
+            "payload": payload,
+            "timeout_seconds": timeout_seconds,
+        })
+        resp.raise_for_status()
+        return RfqResult.from_dict(resp.json())
+
+    def respond_to_rfq(self, rfq_id: str, payload: dict) -> None:
+        """Submit a quote response to an open RFQ."""
+        path = f"/v1/broker/rfq/{rfq_id}/respond"
+        resp = self._authed_request("POST", path, json={"payload": payload})
+        resp.raise_for_status()
+
+    def get_rfq(self, rfq_id: str) -> RfqResult:
+        """Get RFQ status and collected quotes."""
+        path = f"/v1/broker/rfq/{rfq_id}"
+        resp = self._authed_request("GET", path)
+        resp.raise_for_status()
+        return RfqResult.from_dict(resp.json())
+
+    # ── Transaction Tokens ──────────────────────────────────────────
+
+    def request_transaction_token(
+        self,
+        txn_type: str,
+        payload_hash: str,
+        *,
+        session_id: str | None = None,
+        counterparty_agent_id: str | None = None,
+        resource_id: str | None = None,
+        rfq_id: str | None = None,
+        ttl_seconds: int = 60,
+    ) -> dict:
+        """Request a single-use transaction token for a specific operation."""
+        path = "/v1/auth/token/transaction"
+        body: dict[str, Any] = {
+            "agent_id": self._label,
+            "txn_type": txn_type,
+            "payload_hash": payload_hash,
+            "ttl_seconds": ttl_seconds,
+        }
+        if session_id:
+            body["session_id"] = session_id
+        if counterparty_agent_id:
+            body["target_agent_id"] = counterparty_agent_id
+        if resource_id:
+            body["resource_id"] = resource_id
+        if rfq_id:
+            body["rfq_id"] = rfq_id
+        resp = self._authed_request("POST", path, json=body)
+        resp.raise_for_status()
+        return resp.json()
+
+    # ── Notifications ───────────────────────────────────────────────
+
+    def get_notifications(self) -> list[dict]:
+        """Get pending notifications for this agent."""
+        path = "/v1/broker/notifications"
+        resp = self._authed_request("GET", path)
+        resp.raise_for_status()
+        return resp.json()
+
+    # ── Onboarding ──────────────────────────────────────────────────
+
+    @classmethod
+    def join_network(
+        cls,
+        broker_url: str,
+        org_id: str,
+        display_name: str,
+        secret: str,
+        ca_certificate: str,
+        invite_token: str,
+        *,
+        contact_email: str = "",
+        webhook_url: str | None = None,
+        verify_tls: bool = True,
+    ) -> dict:
+        """
+        Request to join the Cullis network as a new organization.
+
+        Requires an invite token generated by the network admin.
+        The request will be reviewed by the network admin after validation.
+        """
+        url = broker_url.rstrip("/") + "/v1/onboarding/join"
+        body: dict[str, Any] = {
+            "org_id": org_id,
+            "display_name": display_name,
+            "secret": secret,
+            "ca_certificate": ca_certificate,
+            "invite_token": invite_token,
+        }
+        if contact_email:
+            body["contact_email"] = contact_email
+        if webhook_url:
+            body["webhook_url"] = webhook_url
+
+        # Lazy import: cullis_sdk.client is mid-loading when this module
+        # is imported at CullisClient definition time, so the module
+        # helper is not yet bound. Defer until the classmethod actually
+        # runs.
+        from cullis_sdk.client import _check_insecure_tls
+        _check_insecure_tls(verify_tls)
+        with httpx.Client(verify=verify_tls) as client:
+            resp = client.post(url, json=body)
+            resp.raise_for_status()
+            return resp.json()
